+++ endpoint/docker.go
@@ -1,4 +1,4 @@
-package actions
+package endpoint
 
 import (
 	"crypto/tls"
@@ -6,31 +6,60 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"net/url"
+	"os/user"
+	"strings"
 	"time"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/samalba/dockerclient"
 )
 
-var (
-	endpointFactory EndpointFactory
-)
-
 func init() {
-	endpointFactory = func(endpointOpts EndpointOptions) (Endpoint, error) {
+}
 
-		tlsConfig, err := getTlsConfig(endpointOpts)
-		if err != nil {
-			return nil, err
-		}
-
-		c, err := dockerclient.NewDockerClient(endpointOpts.Host, tlsConfig)
-		if err != nil {
-			return nil, err
-		}
-
-		return &DockerEndpoint{url: endpointOpts.Host, client: c}, nil
+func NewEndpoint(endpointOpts EndpointOptions) (Endpoint, error) {
+	tlsConfig, err := getTlsConfig(endpointOpts)
+	if err != nil {
+		return nil, err
 	}
+
+	c, err := dockerclient.NewDockerClient(endpointOpts.Host, tlsConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	return &DockerEndpoint{url: endpointOpts.Host, client: c}, nil
+}
+
+type EndpointOptions struct {
+	Host      string
+	TLS       bool
+	TLSVerify bool
+	TLSCaCert string
+	TLSCert   string
+	TLSKey    string
+}
+
+func (eo EndpointOptions) tlsCaCert() string {
+	return resolveHomeDirectory(eo.TLSCaCert)
+}
+
+func (eo EndpointOptions) tlsCert() string {
+	return resolveHomeDirectory(eo.TLSCert)
+}
+
+func (eo EndpointOptions) tlsKey() string {
+	return resolveHomeDirectory(eo.TLSKey)
+}
+
+func resolveHomeDirectory(path string) string {
+	if strings.Contains(path, "~") {
+		usr, _ := user.Current()
+		dir := usr.HomeDir
+		return strings.Replace(path, "~", dir, 1)
+	}
+
+	return path
 }
 
 func getTlsConfig(endpointOpts EndpointOptions) (*tls.Config, error) {
@@ -147,6 +176,10 @@ func (e *DockerEndpoint) ResolveImage(name string) (string, error) {
 	return imageInfo.Id, nil
 }
 
+//func (e *DockerEndpoint) DoRequest(url string, body []byte) ([]byte, error) {
+
+//}
+
 func (e *DockerEndpoint) InspectContainer(name string) (*dockerclient.ContainerInfo, error) {
 	return e.client.InspectContainer(name)
 }
@@ -166,4 +199,61 @@ func translateContainerConfig(cc ContainerConfig) (dockerclient.ContainerConfig,
 
 	err = json.Unmarshal(j, &dcc)
 	return dcc, err
+}
+
+type ContainerConfig struct {
+	dockerclient.ContainerConfig
+	Tty        FromStringOrBool
+	OpenStdin  FromStringOrBool
+	Entrypoint FromStringOrStringSlice
+	// This is used only by the create command
+	HostConfig HostConfig
+}
+
+type HostConfig struct {
+	dockerclient.HostConfig
+	Privileged     FromStringOrBool
+	ReadonlyRootfs FromStringOrBool
+}
+
+type FromStringOrBool struct {
+	Value bool
+}
+
+func (s *FromStringOrBool) UnmarshalJSON(value []byte) error {
+	if value[0] == '"' {
+		var str string
+		if err := json.Unmarshal(value, &str); err != nil {
+			return err
+		}
+		s.Value = (strings.ToLower(str) == "true")
+		return nil
+	}
+
+	return json.Unmarshal(value, &s.Value)
+}
+
+func (s FromStringOrBool) MarshalJSON() ([]byte, error) {
+	return json.Marshal(s.Value)
+}
+
+type FromStringOrStringSlice struct {
+	Value []string
+}
+
+func (s *FromStringOrStringSlice) UnmarshalJSON(value []byte) error {
+	if value[0] == '"' {
+		var str string
+		if err := json.Unmarshal(value, &str); err != nil {
+			return err
+		}
+		s.Value = strings.Split(str, " ")
+		return nil
+	}
+
+	return json.Unmarshal(value, &s.Value)
+}
+
+func (s FromStringOrStringSlice) MarshalJSON() ([]byte, error) {
+	return json.Marshal(s.Value)
 }
